feat: add ButtonPreview component and associated templates for various button schemes
