+++ test/components/previews/pathogen/button_preview/all_schemes.html.erb
