+++ test/components/previews/pathogen/button_preview/leading_visual.html.erb
