+++ test/components/previews/pathogen/button_preview/leading_visual_svg.html.erb
