+++ test/components/previews/pathogen/button_preview/trailing_visual_svg.html.erb
